changement integer en double map

+++ src/Formation.java
@@ -2,9 +2,9 @@ import java.util.HashMap;
 
 public class Formation {
     public String identifiant;
-    public HashMap<String, Integer> mat;
+    public HashMap<String, Double> mat;
 
-    public void addMat(String nom, int coeff){
+    public void addMat(String nom, double coeff){
         if(!this.mat.containsKey(nom))
           this.mat.put(nom, coeff);
         else throw new Error("Impossible d'ajouter une matiere deja existante");
@@ -16,7 +16,7 @@ public class Formation {
         else throw new Error("La matiere n'existe pas");
     }
 
-    public float getCoeff(String nom){
+    public double getCoeff(String nom){
         if(this.mat.containsKey(nom))
             return this.mat.get(nom);
         throw new Error("Nom de matiere inexistant");
